Bugfix for expandable database rows switching foreignKeys (See detailed notes)
I cannot replicate this in unit tests, but I can replicate it in our application that uses Expandable.
We have "Member" as a model with "MembersField" as an expandable.

I create a controller that only does:
$this->Member->save(array('id' => 1, 'job_credit_total' => $total));

The first time I run it, the correct members field entry is there:
+---------+---------------------+---------------------+-----------+------------------+-------+
| id      | created             | modified            | member_id | key              | value |
+---------+---------------------+---------------------+-----------+------------------+-------+
| 1353409 | 2014-01-17 10:32:27 | 2014-01-17 10:32:30 |         1 | job_credit_total | 10    |
+---------+---------------------+---------------------+-----------+------------------+-------+

The second time I run it, the member_id is changed to the currently logged in user:
+---------+---------------------+---------------------+-----------+------------------+-------+
| id      | created             | modified            | member_id | key              | value |
+---------+---------------------+---------------------+-----------+------------------+-------+
| 1353409 | 2014-01-17 10:32:27 | 2014-01-17 10:33:35 |    120854 | job_credit_total | 10    |
+---------+---------------------+---------------------+-----------+------------------+-------+

There is some sort of interaction between another query our site is doing in regards to
the currently logged in user and MembersField that is changing the state of the $Member->MembersField
object.   I cannot pin it down.

However, moving this code to a seemingly redundant location fixes the problem.